requirements.txt
# Core ML and Backend Dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
scikit-learn>=1.3.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# FastAPI Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
sqlite3
psycopg2-binary>=2.9.0  # For PostgreSQL

# Authentication & Security
email-validator>=2.0.0
cryptography>=41.0.0

# Data Processing and Analysis
pandas>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning Enhancement
pytorch-grad-cam>=1.4.6
torchcam>=0.4.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Image Processing
imageio>=2.31.0
albumentations>=1.3.0

# Dataset Management
datasets>=2.14.0
huggingface-hub>=0.16.0
transformers>=4.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
click>=8.1.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.5.0

# Deployment
gunicorn>=21.0.0
redis>=4.6.0
celery>=5.3.0

# Optional: GPU acceleration
# torch-audio>=2.0.0  # if audio features needed
# torchtext>=0.15.0   # if text processing needed

# Production extras
psutil>=5.9.0
cryptography>=41.0.0
httpx>=0.24.0